+++ docs/src/js/appRoutes.jsx
@@ -1,11 +1,9 @@
 'use strict'
 
-import {Route, Router} from 'react-router'
+import {Route, Router, hashHistory} from 'react-router'
 import Master from './pages/master.jsx'
 import Home from './pages/home.jsx'
-import createHistory from 'history/lib/createHashHistory'
 
-const history = createHistory({ queryKey: false })
 let menulist = []
 
 require('./menuList').forEach(function (menu, index) {
@@ -17,7 +15,7 @@ require('./menuList').forEach(function (menu, index) {
 })
 
 const AppRoutes = (
-  <Router onUpdate={() => window.scrollTo(0, 0)} history={history}>
+  <Router onUpdate={() => window.scrollTo(0, 0)} history={hashHistory}>
     <Route path="/" indexRoute={{component: Home}} component={Master}>
       <Route path="/home" component={Home} />
       {menulist}