+++ docs/src/js/pages/switch.jsx
@@ -1,0 +1,68 @@
+'use strict'
+
+import React from 'react'
+import Code from '../Code'
+import Example from '../Example'
+const {Switch} = global.uiRequire()
+
+module.exports = class extends React.Component {
+  render () {
+    return (
+      <div>
+        <div className="header">
+          <h1>Switch</h1>
+          <h2>复选框</h2>
+        </div>
+
+        <div className="content">
+
+          <Code>
+{`<Switch
+  className={string}  // class
+  text="string"       // 显示的文字信息
+  value={bool}         // 传入校验是否选中的值，只用来做校验，选中后实际返回值是defaultValue
+  checked={bool}      // 是否选中，默认为 false
+  readOnly={bool}     // 是否只读，默认为 false
+  onChange={function} // 状态改变回调事件
+/>`}
+          </Code>
+
+          <h2 className="subhead">Example</h2>
+          <Example>
+<Switch />
+          </Example>
+
+          <h2 className="subhead">Text</h2>
+          <Example>
+<Switch text="Yes|No" />
+          </Example>
+
+          <h2 className="subhead">Round</h2>
+          <Example>
+<Switch round text="Yes|No" />
+          </Example>
+
+          <h2 className="subhead">Small</h2>
+          <Example>
+<Switch size="small" />
+<Switch size="small" round />
+          </Example>
+
+          <h2 className="subhead">Large</h2>
+          <Example>
+<Switch size="large" />
+<Switch size="large" round />
+          </Example>
+
+          <h2 className="subhead">Readonly</h2>
+          <Example>
+<Switch checked readOnly />
+<Switch round checked readOnly />
+<Switch readOnly />
+<Switch round readOnly />
+          </Example>
+        </div>
+      </div>
+    )
+  }
+}