+++ src/index.js
@@ -17,6 +17,7 @@ export { default as Select } from './Select'
 export { default as Tree } from './Tree'
 export { default as Upload } from './Upload'
 export { default as Button } from './Button'
+export { default as Switch } from './Switch'
 
 export { default as FormControl } from './FormControl'
 export { default as FormSubmit } from './FormSubmit'