+++ src/themes/index.js
@@ -1,8 +1,0 @@
-'use strict'
-
-export let THEME = 'pure'
-
-export function requireCss (pack) {
-  console.warn(`${pack} wait a minute...`)
-  // require(`./pure/${pack}.less`);
-};